@@ -3,14 +3,14 @@ import java.nio.file.Paths
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.4.21'
-    ext.protobufVersion = '0.8.14'
+    ext.kotlin_version = '1.6.21'
+    ext.protobufVersion = '0.9.0'
     repositories {
         google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.1.1'
+        classpath 'com.android.tools.build:gradle:8.1.0'
         classpath "com.google.protobuf:protobuf-gradle-plugin:$protobufVersion"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
 
