@@ -16,7 +16,7 @@
     import OpenGLES
 #elseif os(Android)
     import Glibc
-    import GL.ES3
+    import ndk.GLES3
 #endif
 
 
