@@ -17,7 +17,7 @@ import OpenGL
 import OpenGLES
 #elseif os(Android)
 import Glibc
-import GL.ES3
+import ndk.GLES3
 #endif
 
 func BUFFER_OFFSET(_ i: Int) -> UnsafeRawPointer? {
